[API] move route
