+++ api/lib/application/admin-members/index.js
@@ -37,34 +37,6 @@ exports.register = async function (server) {
     },
 
     {
-      method: 'POST',
-      path: '/api/admin/admin-members',
-      config: {
-        pre: [{ method: securityPreHandlers.checkAdminMemberHasRoleSuperAdmin }],
-        validate: {
-          payload: Joi.object({
-            data: Joi.object({
-              attributes: Joi.object({
-                email: Joi.string().email().required(),
-                role: Joi.string().valid('SUPER_ADMIN', 'SUPPORT', 'METIER', 'CERTIF').required(),
-              }),
-            }),
-          }),
-          options: {
-            allowUnknown: true,
-          },
-        },
-        handler: adminMemberController.saveAdminMember,
-        notes: [
-          "- Cette route est restreinte aux utilisateurs ayant les droits d'accès\n" +
-            '- Elle permet de donner un accès à Pix Admin à un nouveau membre\n' +
-            'ou à réactiver un membre désactivé',
-        ],
-        tags: ['api', 'admin', 'admin-members'],
-      },
-    },
-
-    {
       method: 'PATCH',
       path: '/api/admin/admin-members/{id}',
       config: {
