+++ api/tests/unit/application/admin-members/index_test.js
@@ -58,142 +58,6 @@ describe('Unit | Application | Router | admin-members-router', function () {
     });
   });
 
-  describe('POST /api/admin/admin-members', function () {
-    context('when user has role "SUPER_ADMIN"', function () {
-      it('should return a response with an HTTP status code 201', async function () {
-        // given
-        const adminMember = domainBuilder.buildAdminMember();
-        sinon
-          .stub(adminMemberController, 'saveAdminMember')
-          .callsFake((request, h) => h.response(adminMember).created());
-        sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').returns(true);
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(adminMembersRouter);
-
-        // when
-        const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-          data: { type: 'admin-members', attributes: { role: ROLES.SUPER_ADMIN, email: 'fire.bot@example.net' } },
-        });
-
-        // then
-        expect(statusCode).to.equal(201);
-      });
-
-      context('when role attribute is missing', function () {
-        it('should return a response with an HTTP status code 400', async function () {
-          // given
-          sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').returns(true);
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(adminMembersRouter);
-
-          // when
-          const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-            data: { type: 'admin-members', attributes: { email: 'fire.bot@example.net' } },
-          });
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-
-      context('when role attribute is not valid', function () {
-        it('should return a response with an HTTP status code 400', async function () {
-          // given
-          const adminMember = domainBuilder.buildAdminMember();
-          sinon.stub(adminMemberController, 'saveAdminMember').returns(adminMember);
-          sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').returns(true);
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(adminMembersRouter);
-
-          // when
-          const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-            data: { type: 'admin-members', attributes: { role: 'My-Role', email: 'fire.bot@example.net' } },
-          });
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-
-      context('when email attribute is missing', function () {
-        it('should return a response with an HTTP status code 400', async function () {
-          // given
-          const adminMember = domainBuilder.buildAdminMember();
-          sinon.stub(adminMemberController, 'saveAdminMember').returns(adminMember);
-          sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').returns(true);
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(adminMembersRouter);
-
-          // when
-          const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-            data: { type: 'admin-members', attributes: { role: ROLES.SUPER_ADMIN } },
-          });
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-
-      context('when email attribute is not valid', function () {
-        it('should return a response with an HTTP status code 400', async function () {
-          // given
-          const adminMember = domainBuilder.buildAdminMember();
-          sinon.stub(adminMemberController, 'saveAdminMember').returns(adminMember);
-          sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').returns(true);
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(adminMembersRouter);
-
-          // when
-          const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-            data: { type: 'admin-members', attributes: { role: ROLES.SUPER_ADMIN, email: 'my-invalid-email' } },
-          });
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-
-      context('when all attributes are missing', function () {
-        it('should return a response with an HTTP status code 400', async function () {
-          // given
-          const adminMember = domainBuilder.buildAdminMember();
-          sinon.stub(adminMemberController, 'saveAdminMember').returns(adminMember);
-          sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').returns(true);
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(adminMembersRouter);
-
-          // when
-          const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-            data: { type: 'admin-members', attributes: {} },
-          });
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-    });
-
-    context('when user does not have role "SUPER_ADMIN"', function () {
-      it('should return a response with an HTTP status code 403', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-          .callsFake((request, h) => h.response().code(403).takeover());
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(adminMembersRouter);
-
-        // when
-        const { statusCode } = await httpTestServer.request('POST', '/api/admin/admin-members', {
-          data: { type: 'admin-members', attributes: { role: ROLES.SUPER_ADMIN, email: 'aqua.bot@example.net' } },
-        });
-
-        // then
-        expect(statusCode).to.equal(403);
-        expect(securityPreHandlers.checkAdminMemberHasRoleSuperAdmin).to.have.be.called;
-      });
-    });
-  });
-
   describe('PATCH /api/admin/admin-members/{id}', function () {
     describe('when user has role "SUPER_ADMIN"', function () {
       it('should return a response with an HTTP status code 200', async function () {
